\name{highschool}
\alias{highschool}
\docType{data}
\title{Highschool data on alcohol, cigarette and marijuana use for
  highschool seniors}
\description{
Data from a 1992 survey by the Wright State University School of
Medecine and the United Health Services in Dayton, Ohio. The survey
asked 2276 students in their final year of highschool in a nonurban area
near Dayton, Ohio, whether they had ever used alcohol, cigarettes, or marijuana.
}

\usage{data(highschool)
}

\format{
  A data frame of 8 observations on the variables:
  \describe{
\item{alcohol}{ a factor vector with components "yes" or "no".}
\item{cigarette}{ a factor vector with components "yes" or "no".}
\item{marijuana}{ a factor vector with components "yes" or "no".}
\item{count}{ a numeric vector of frequencies.}
 }
}

\references{
Agresti A. (2002). Categorical data analysis, Wiley, p. 322.
}

\examples{
data(highschool)
}
\keyword{datasets}
