% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndependenceTests-package.R
\docType{package}
\name{IndependenceTests-package}
\alias{IndependenceTests}
\alias{IndependenceTests-package}
\title{IndependenceTests: Non-Parametric Tests of Independence Between Random Vectors}
\description{
Functions for non-parametric tests of independence (mutual or serial) between some quantitative random vectors, as described in Bilodeau M. and Lafaye de Micheaux P. (2009) \doi{10.1016/j.jspi.2008.11.006}, in Beran R., Bilodeau M. and Lafaye de Micheaux P. (2007) \doi{10.1016/j.jmva.2007.01.009} and in Fan Y., Lafaye de Micheaux P., Penev S. and Salopek D. (2017) \doi{10.1016/j.jmva.2016.09.014}.
}
\author{
\strong{Maintainer}: Pierre Lafaye De Micheaux \email{lafaye@unsw.edu.au}

Authors:
\itemize{
  \item Martin Bilodeau \email{bilodeau@dms.umontreal.ca}
  \item Yanan Fan \email{y.fan@unsw.edu.au}
  \item Spiridon Penev \email{s.penev@unsw.edu.au}
  \item Donna Salopek \email{dm.salopek@unsw.edu.au}
}

Other contributors:
\itemize{
  \item Cleve Moler (LAPACK/BLAS routines in src) [copyright holder]
  \item Jack Dongarra (LAPACK/BLAS routines in src) [copyright holder]
  \item Richard Hanson (LAPACK/BLAS routines in src) [copyright holder]
  \item Sven Hammarling (LAPACK/BLAS routines in src) [copyright holder]
  \item Jeremy Du Croz (LAPACK/BLAS routines in src) [copyright holder]
}

}
\keyword{internal}
