% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdcov.R
\name{mdcov}
\alias{mdcov}
\title{Computation of the test statistic}
\usage{
mdcov(
  X,
  vecd,
  a = 1,
  weight.choice = 1,
  N = 200,
  cubature = FALSE,
  K = 100,
  epsrel = 10^-6,
  norming = TRUE,
  thresh.eigen = 10^-8,
  estim.a = FALSE,
  Cpp = TRUE,
  pval.comp = TRUE
)
}
\arguments{
\item{X}{matrix of size n x q}

\item{vecd}{vector of length p giving the sizes q_1,...,q_p of each one of the p subvectors}

\item{a}{real number (parameter in the w() weight function)}

\item{weight.choice}{weight.choice of the weight function as described in our paper.}

\item{N}{...}

\item{cubature}{...}

\item{K}{...}

\item{epsrel}{...}

\item{norming}{...}

\item{thresh.eigen}{...}

\item{estim.a}{...}

\item{Cpp}{...}

\item{pval.comp}{...}
}
\description{
Computation of the test statistic
}
