% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependogram.R
\name{dependogram}
\alias{dependogram}
\title{Title}
\usage{
dependogram(
  X,
  vecd.or.p,
  N = 10,
  B = 2000,
  alpha = 0.05,
  display = TRUE,
  graphics = TRUE,
  nbclus = 1
)
}
\arguments{
\item{X}{}

\item{vecd.or.p}{}

\item{N}{}

\item{B}{}

\item{alpha}{}

\item{display}{}

\item{graphics}{}

\item{nbclus}{}
}
\description{
Title
}
